==> 3.3 VS Topics test and features.ipynb <==
import pandas as pd
from datetime import datetime
import plotly.express as px
import matplotlib.pyplot as plt
import numpy as np
from wordcloud import WordCloud
import plotly.graph_objects as go

import nltk
nltk.download('wordnet')
from nltk.stem import WordNetLemmatizer
from nltk.corpus import stopwords
nltk.download('stopwords')
from nltk.corpus import stopwords
stop_words = set(stopwords.words("english"))
from nltk import FreqDist
from nltk import ngrams

from typing import Any, Optional

from bertopic import BERTopic
#from plotly.offline import init_notebook_mode
#
#init_notebook_mode(connected=True)

import warnings
warnings.filterwarnings("ignore")

def clean_text(columna_datos):
    pattern = r'''(?x)                 # set flag to allow verbose regexps
              (?:[A-Z]\.)+         # abbreviations, e.g. U.S.A.
              | \w+(?:-\w+)*       # words with optional internal hyphens
              | \$?\d+(?:\.\d+)?%? # currency and percentages, e.g. $12.40, 82%
              | \.\.\.             # ellipsis
              | [][.,;"'?():-_`]   # these are separate tokens; includes ], [
    '''
    textos = columna_datos

    # Tokenizar cada fila
    tokens = [nltk.regexp_tokenize(texto, pattern) for texto in textos]

    columna_tokens = pd.DataFrame()
    columna_tokens["Tokens"] = tokens

    # Lematizador
    lemmatizer = WordNetLemmatizer()
    columna_tokens["Tokens"] = columna_tokens["Tokens"].apply(lambda tokens: [lemmatizer.lemmatize(token) for token in tokens])

    # Eliminar stopwords:
    columna_tokens["Tokens"] = columna_tokens["Tokens"].apply(lambda tokens: [token for token in tokens if token.lower() not in stop_words])

    # Eliminar carácteres especiales y palabras cortas:
    columna_tokens["Tokens"] = columna_tokens["Tokens"].apply(lambda tokens: [token for token in tokens if token.isalnum() and len(token) >=3])

    # Convertir todo a minúsculas:
    columna_tokens["Tokens"] = columna_tokens["Tokens"].apply(lambda tokens: [token.lower() for token in tokens])

    # Convertir todo a texto:
    columna_tokens["Tokens"] = columna_tokens["Tokens"].apply(lambda tokens: ' '.join([token.lower() for token in tokens]))

    return columna_tokens

data = pd.read_csv("./Datos ecopetrol/Datos_ecopetrol_semanales_divididos.csv")
data["Date"] = pd.to_datetime(data["Date"])
data = data.sort_values(by="Date")
data["text"] = clean_text(data["Title"])

train_weekly = data.loc[data["Split"] == "Train"]
val_weekly = data.loc[data["Split"] == "Validation"]
test_weekly = data.loc[data["Split"] == "Test"]

topic_model = BERTopic.load("./Modelos/bertopic_model")

%%capture
assigned_topics = []
for i in test_weekly["text"]:
    topics_pred, probs_pred = topic_model.transform(i)
    assigned_topics.extend(topics_pred)
test_weekly["Topic"] = assigned_topics
test_weekly.to_csv("./Datos ecopetrol/Datos_ecopetrol_topicos_test.csv", index=False)

train_weekly = pd.read_csv("./Datos ecopetrol/Datos_ecopetrol_topicos_train.csv")
val_weekly = pd.read_csv("./Datos ecopetrol/Datos_ecopetrol_topicos_validation.csv")
test_weekly = pd.read_csv("./Datos ecopetrol/Datos_ecopetrol_topicos_test.csv")

data_with_topics = pd.concat([train_weekly,val_weekly,test_weekly], ignore_index=True)

data_with_topics = data_with_topics.drop(columns="Unnamed: 0")

data_with_topics.to_csv("./Datos ecopetrol/Datos_ecopetrol_topicos.csv", index=False)

data = data_with_topics.copy()

data["Date"] = pd.to_datetime(data["Date"])

topics = data[["Date","Topic"]]

topics_var = topics['Topic'].unique()
for topic in topics_var:
    topics[f'Topic_{topic}_appeared'] = topics['Date'].isin(topics[topics['Topic'] == topic]['Date']).astype(int)

topics = topics.groupby('Date').agg({
    **{f'Topic_{topic}_appeared': 'max' for topic in topics_var}
}).reset_index()

daily_topics = topics.groupby(topics['Date'].dt.to_period("D")).agg({
    **{f'Topic_{topic}_appeared': 'max' for topic in topics_var}
}).reset_index()

week_topics = topics.groupby(topics['Date'].dt.to_period("W")).agg({
    **{f'Topic_{topic}_appeared': 'max' for topic in topics_var}
}).reset_index()

week_topics["Date"] = pd.to_datetime(week_topics["Date"].astype(str).str.split('/', expand=True)[1])

week_topics.to_csv("./Datos ecopetrol/Datos_ecopetrol_topicos_modelos.csv")

# Para artículo

import pandas as pd

df = pd.read_csv("./Datos ecopetrol/Datos_ecopetrol_topicos_modelos.csv")
df = df.drop(columns="Unnamed: 0")
df

df = {
    'Date': ['1/07/2012', '8/07/2012', '15/07/2012', '22/07/2012', '29/07/2012'],
    'Topic_1': [1, 0, 1, 1, 0],
    'Topic_2': [0, 1, 0, 0, 0],
    'Topic_3': [0, 0, 1, 0, 1],
    'Topic_n': [0, 1, 0, 1, 0]
}

df = pd.DataFrame(data)

latex_table = df.to_latex(index=False)

latex_table
