==> 5.1 Feature selection.ipynb <==
# Libraries

import matplotlib.pyplot as plt
import numpy as np
import pandas as pd
import seaborn as sns

from sklearn.feature_selection import RFE
from sklearn.feature_selection import RFECV
from sklearn.feature_selection import SelectKBest, chi2, f_classif, f_regression 

import warnings
warnings.filterwarnings("ignore")

# Data

data = pd.read_csv("./Datos ecopetrol/DATOS_PREDICTIVOS_EC.csv")
data['Date'] = pd.to_datetime(data['Date'])
data = data.sort_values(by='Date')
data = data.set_index("Date")
data["close1"] = data["Close"].shift(-1)
column_to_insert = data['close1']
data.drop(columns=['close1'], inplace=True)
data.insert(0, 'close1', column_to_insert)

data = data.drop(columns=["Close", "Open","diff","trend","open1"])

topic_columns = [col for col in data.columns if 'Topic' in col]
data[topic_columns] = data[topic_columns].apply(pd.to_numeric, errors='coerce').astype('Int64')

data=data.dropna()

# Get relevant features for only train data
date1 = "2020-07-12"
data = data.loc[data.index<=date1]

trend1 = data.drop(columns=["trend2","trend3","trend4","diff1","diff2","diff3","diff4","close1","close2","close3","close4"])
trend2 = data.drop(columns=["trend1","trend3","trend4","diff1","diff2","diff3","diff4","close1","close2","close3","close4"])
trend3 = data.drop(columns=["trend1","trend2","trend4","diff1","diff2","diff3","diff4","close1","close2","close3","close4"])
trend4 = data.drop(columns=["trend1","trend2","trend3","diff1","diff2","diff3","diff4","close1","close2","close3","close4"])

diff1 = data.drop(columns=["trend1","trend2","trend3","trend4","diff2","diff3","diff4","close1","close2","close3","close4"])
diff2 = data.drop(columns=["trend1","trend2","trend3","trend4","diff1","diff3","diff4","close1","close2","close3","close4"])
diff3 = data.drop(columns=["trend1","trend2","trend3","trend4","diff1","diff2","diff4","close1","close2","close3","close4"])
diff4 = data.drop(columns=["trend1","trend2","trend3","trend4","diff1","diff2","diff3","close1","close2","close3","close4"])

# SelectKBest for trends

## Categorical

dataframes = [trend1,trend2,trend3,trend4]

topic_columns = [col for col in data.columns if 'Topic' in col]

threshold = 1

objectives = []
final_list = []

for i in dataframes:
    X_train = i[topic_columns]
    y_train = i.iloc[:,:1]

    select_feature = SelectKBest(chi2, k=10).fit(X_train, y_train)

    selected_features_df = pd.DataFrame({'Feature':list(X_train.columns),
                                     'Scores':select_feature.scores_})
    
    features_list = selected_features_df[selected_features_df['Scores'] > threshold]['Feature'].tolist()

    objectives.append(i.iloc[:,:1].columns[0])
    final_list.append(features_list)

    print(i.iloc[:,:1].columns[0], features_list)

cat_features = pd.DataFrame({"Objective":objectives,"Variables":final_list})

## Escalar

dataframes = [trend1,trend2,trend3,trend4]

topic_columns = [col for col in data.columns if 'Topic' in col]

threshold = 1

objectives = []
final_list = []

for i in dataframes:

    X_train = i[i.columns.difference(topic_columns)]
    X_train = X_train.drop(columns=i.iloc[:,:1].columns[0])
    y_train = i.iloc[:,:1]

    select_feature = SelectKBest(f_classif, k=10).fit(X_train, y_train)

    selected_features_df = pd.DataFrame({'Feature':list(X_train.columns),
                                     'Scores':select_feature.scores_})
    
    features_list = selected_features_df[selected_features_df['Scores'] > threshold]['Feature'].tolist()

    objectives.append(i.iloc[:,:1].columns[0])
    final_list.append(features_list)

    print(i.iloc[:,:1].columns[0], features_list)

esc_features = pd.DataFrame({"Objective":objectives,"Variables":final_list})

# SelectKBest for volatility


## Categorical

dataframes = [diff1,diff2,diff3,diff4]

topic_columns = [col for col in data.columns if 'Topic' in col]

threshold = 1

objectives = []
final_list = []

for i in dataframes:
    X_train = i[topic_columns]
    y_train = i.iloc[:,:1]

    select_feature = SelectKBest(f_regression, k=10).fit(X_train, y_train)

    selected_features_df = pd.DataFrame({'Feature':list(X_train.columns),
                                     'Scores':select_feature.scores_})
    
    features_list = selected_features_df[selected_features_df['Scores'] > threshold]['Feature'].tolist()

    objectives.append(i.iloc[:,:1].columns[0])
    final_list.append(features_list)

    print(i.iloc[:,:1].columns[0], features_list)

cat_features = pd.DataFrame({"Objective":objectives,"Variables":final_list})

## Escalar

dataframes = [diff1,diff2,diff3,diff4]

topic_columns = [col for col in data.columns if 'Topic' in col]

threshold = 1

objectives = []
final_list = []

for i in dataframes:

    X_train = i[i.columns.difference(topic_columns)]
    X_train = X_train.drop(columns=i.iloc[:,:1].columns[0])
    y_train = i.iloc[:,:1]

    select_feature = SelectKBest(f_regression, k=10).fit(X_train, y_train)

    selected_features_df = pd.DataFrame({'Feature':list(X_train.columns),
                                     'Scores':select_feature.scores_})
    
    features_list = selected_features_df[selected_features_df['Scores'] > threshold]['Feature'].tolist()

    objectives.append(i.iloc[:,:1].columns[0])
    final_list.append(features_list)

    print(i.iloc[:,:1].columns[0], features_list)
